[4주차] fix Error - index js
