--- 윤다빈/freelec-springboot2-webservice/.gitignore
@@ -1,2 +1,0 @@
-.gradle
-.idea
