+++ 윤다빈/freelec-springboot2-webservice/.gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Sat Nov 05 18:31:54 KST 2022
+gradle.version=4.10.2
