--- 윤다빈/freelec-springboot2-webservice/.idea/.gitignore
@@ -1,3 +1,0 @@
-# 디폴트 무시된 파일
-/shelf/
-/workspace.xml
