+++ 윤다빈/freelec-springboot2-webservice/src/main/java/com/jojoldu/book/springboot/domain/posts/Posts.java
@@ -12,8 +12,8 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 @Getter
-@NoArgsConstructor     // @Getter 와 @NoArgsConstructor(기본 생성자 자동 추가, public Post(){}의 효과) 는 롬복의 어노테이션으로 필수 아님
-@Entity                // JPA의 어노테이션으로 필수라 클래스에 가까이 뒀음
+@NoArgsConstructor // @Getter 와 @NoArgsConstructor(기본 생성자 자동 추가, public Post(){}의 효과) 는 롬복의 어노테이션으로 필수 아님
+@Entity            // JPA의 어노테이션으로 필수라 클래스에 가까이 뒀음
 public class Posts extends BaseTimeEntity {
 
     @Id
@@ -28,7 +28,7 @@ public class Posts extends BaseTimeEntity {
 
     private String author;
 
-    @Builder  // 해당 클래스의 빌더 패턴 클래스 생성
+    @Builder // 해당 클래스의 빌더 패턴 클래스 생성
     public Posts(String title, String content, String author) {
         this.title = title;
         this.content = content;
