+++ 윤다빈/freelec-springboot2-webservice/src/main/java/com/jojoldu/book/springboot/service/posts/PostsService.java
@@ -20,7 +20,6 @@ public class PostsService {
 
     @Transactional
     public Long save(PostsSaveRequestDto requestDto) {
-
         return postsRepository.save(requestDto.toEntity()).getId();
     }
 
@@ -28,30 +27,32 @@ public class PostsService {
     public Long update(Long id, PostsUpdateRequestDto requestDto) {
         Posts posts = postsRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("해당 사용자가 없습니다. id=" + id));
+
         posts.update(requestDto.getTitle(), requestDto.getContent());
+
         return id;
     }
 
-    @Transactional(readOnly = true)
-    public PostsResponseDto findById(Long id) {
-        Posts entity = postsRepository.findById(id)
+    @Transactional
+    public void delete (Long id) {
+        Posts posts = postsRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("해당 사용자가 없습니다. id=" + id));
-        return new PostsResponseDto(entity);
+
+        postsRepository.delete(posts);
     }
 
     @Transactional(readOnly = true) // readOnly=true 옵션으로 조회 속도를 개선시킴
-    public List<PostsListResponseDto> findAllDesc(){
+    public PostsResponseDto findById(Long id) {
+        Posts entity = postsRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("해당 사용자가 없습니다. id=" + id));
+
+        return new PostsResponseDto(entity);
+    }
+
+    @Transactional(readOnly = true)
+    public List<PostsListResponseDto> findAllDesc() {
         return postsRepository.findAllDesc().stream()
                 .map(PostsListResponseDto::new)
                 .collect(Collectors.toList());
     }
-
-    @Transactional
-    public void delete(Long id){
-        Posts posts = postsRepository.findById(id)
-                .orElseThrow(()-> new IllegalArgumentException("해당 게시글이 없습니다. id=" + id));
-        postsRepository.delete(posts);
-    }
-
 }
-
