+++ 윤다빈/freelec-springboot2-webservice/src/main/java/com/jojoldu/book/springboot/web/HelloController.java
@@ -14,7 +14,8 @@ public class HelloController {
     }
 
     @GetMapping("/hello/dto")
-    public HelloResponseDto helloDto(@RequestParam("name") String name, @RequestParam("amount") int amount){
+    public HelloResponseDto helloDto(@RequestParam("name") String name, @RequestParam("amount") int amount) {
         return new HelloResponseDto(name, amount);
     }
+
 }
