+++ 윤다빈/freelec-springboot2-webservice/src/main/java/com/jojoldu/book/springboot/web/dto/PostsListResponseDto.java
@@ -3,17 +3,16 @@ package com.jojoldu.book.springboot.web.dto;
 import com.jojoldu.book.springboot.domain.posts.Posts;
 import lombok.Getter;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Getter
 public class PostsListResponseDto {
     private Long id;
     private String title;
-    private  String author;
+    private String author;
     private LocalDateTime modifiedDate;
 
-    public PostsListResponseDto(Posts entity){
+    public PostsListResponseDto(Posts entity) {
         this.id = entity.getId();
         this.title = entity.getTitle();
         this.author = entity.getAuthor();
