+++ 윤다빈/freelec-springboot2-webservice/src/main/java/com/jojoldu/book/springboot/web/dto/PostsResponseDto.java
@@ -11,7 +11,7 @@ public class PostsResponseDto {
     private String content;
     private String author;
 
-    public PostsResponseDto(Posts entity){
+    public PostsResponseDto(Posts entity) {
         this.id = entity.getId();
         this.title = entity.getTitle();
         this.content = entity.getContent();
