+++ 윤다빈/freelec-springboot2-webservice/src/main/resources/application.properties
@@ -1,3 +1,5 @@
-spring.jpa.show_sql = true
+spring.profiles.include=real-oauth
+spring.jpa.show_sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
 spring.h2.console.enabled=true
+spring.session.store-type=jdbc
