+++ 윤다빈/freelec-springboot2-webservice/src/main/resources/templates/layout/footer.mustache
@@ -5,4 +5,3 @@
 <script src="/js/app/index.js"></script>
 </body>
 </html>
-
