+++ 윤다빈/freelec-springboot2-webservice/src/test/java/com/jojoldu/book/springboot/domain/posts/PostsRepositoryTest.java
@@ -1,7 +1,5 @@
 package com.jojoldu.book.springboot.domain.posts;
 
-import com.jojoldu.book.springboot.domain.posts.Posts;
-import com.jojoldu.book.springboot.domain.posts.PostsRepository;
 import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
