+++ 윤다빈/freelec-springboot2-webservice/src/test/java/com/jojoldu/book/springboot/web/IndexControllerTest.java
@@ -12,18 +12,18 @@ import static org.springframework.boot.test.context.SpringBootTest.WebEnvironmen
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = RANDOM_PORT)
-public class indexControllerTest {
+public class IndexControllerTest {
+
     @Autowired
     private TestRestTemplate restTemplate;
 
     @Test
     public void 메인페이지_로딩() {
-        // when
+        //when
         String body = this.restTemplate.getForObject("/", String.class);
 
-        // then
+        //then
         // HTML은 규치있는 문자열이므로, /로 호출했을때, 일부 문자열이 포함되어있는지만 비교하면 됨
-        assertThat(body).contains("스프링 부트로 시작하는 웹 서비스");
+        assertThat(body).contains("스프링부트로 시작하는 웹 서비스");
     }
 }
-
