+++ 윤다빈/freelec-springboot2-webservice/src/test/resources/application.properties
@@ -1,0 +1,10 @@
+spring.jpa.show_sql=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
+spring.h2.console.enabled=true
+spring.session.store-type=jdbc
+
+# Test OAuth
+
+spring.security.oauth2.client.registration.google.client-id=test
+spring.security.oauth2.client.registration.google.client-secret=test
+spring.security.oauth2.client.registration.google.scope=profile,email
